+++ sdk/core/azure-core/src/private/environment_log_level_listener.hpp
@@ -25,7 +25,7 @@ namespace Azure { namespace Core { namespace Diagnostics { namespace _detail {
 
   public:
     static Logger::Level GetLogLevel(Logger::Level defaultValue);
-    static Logger::Listener GetLogListener();
+    static std::function<void(Logger::Level level, std::string const& message)> GetLogListener();
   };
 
 #if (defined(WINAPI_PARTITION_DESKTOP) && !WINAPI_PARTITION_DESKTOP) // See azure/core/platform.hpp
@@ -35,6 +35,10 @@ namespace Azure { namespace Core { namespace Diagnostics { namespace _detail {
     return defaultValue;
   }
 
-  inline Logger::Listener EnvironmentLogLevelListener::GetLogListener() { return nullptr; }
+  inline std::function<void(Logger::Level level, std::string const& message)>
+  EnvironmentLogLevelListener::GetLogListener()
+  {
+    return nullptr;
+  }
 #endif
 }}}} // namespace Azure::Core::Diagnostics::_detail