+++ sdk/core/azure-core/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Removed `Context::GetApplicationContext()` in favor of a new static data member `Context::ApplicationContext`.
 - Renamed `Request::IsDownloadViaStream()` to `ShouldBufferResponse()`.
 - Removed the `Azure::Core::Http::Request` ctor overload that takes both a `bodyStream` and a `bufferedDownload` boolean since it is not useful.
+- Removed the `Azure::Core::Diagnostics::Logger::Listener` typedef.
 
 ### Bug Fixes